rockchip/rk3399: sdram: fix data training function
1. override write leveing value to 0x200.
When the wrdqs slave delay is changed to 0x200, the phase between dqs
and the clock is 0 degrees. The pcb layout can make sure tDQSS timing
is smaller than 0.25tck, so this value is useful for both higher and
lower frequencies.

2. disable read leveing for LPDDR3.
Read leveing result is unreliable,the value is not in the middle of
read eye. So disable read leveing and fix the read DQSn slave delay
setting for DQn to 0x080(1/4 cycle delay of the input signal).

Check by shmoo read eye and stability test, fix to 0x80 is better.

BUG=None
BRANCH=None
TEST=Boot from kevin

Change-Id: I2a5d40c0348449b2a7c609c1db65da4ed5f1c09f
Signed-off-by: Lin Huang <hl@rock-chips.com>
Signed-off-by: Jeff Chen <cym@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/396598
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>